refactor(channels): move non-trivial capabilities.h function bodies to capabilities.cpp.hpp (#3186)
CodeRabbit feedback on #3197: function definitions in `.h` violate the
repo rule for the channels subsystem. Move out-of-line:

- `PinSet::fixed(i16)`
- `PinSet::accepts(i16) const`
- `PinSet::acceptsAll(const PinBitset&) const`
- `DriverCapabilities::DriverCapabilities()`
- `ChannelRequest::singlePin(...)`

Kept in the header (constexpr-required or template-required):

- `FreqRange` / `NanosRange` factories and member queries (single-return
  constexpr — must be in header for constexpr-evaluable usage in
  `static constexpr PinGroup` initializers).
- `PinSet` constructors (member-init only, no body).
- `PinSet::none()` / `anyOutput()` (single-return constexpr).
- `PinSet::allowlist<N>(...)` (template — must be in header).
- `ClocklessTimingCapability::unspecified()` / `isUnspecified()` (single
  return constexpr).
- `PinGroup::PinGroup()` (member-init only).
- `ChipsetClocklessTiming::unspecified()` (single-return constexpr).
- `ChannelRequest::ChannelRequest()` (member-init only).
- `isYes(HandleResult)` (single-return constexpr free function).

New `capabilities.cpp.hpp` is wired into `_build.cpp.hpp` so the unity
build provides single-definition external linkage; test TUs that link
against fastled.dll resolve the symbols normally.

All 329 C++ unit tests still pass.
